Add k8s api node controller
This commit adds the node controller that watches for updates to nodes.
As notifications are received they are then queued up to a work queue
for further processing.

The image built uses the inClusterConfig to find and connect to the
k8s api. service accounts and roles need to be created to allow
the application to access the different k8s resources.

yaml files are included for:
- creating a stargazer pod
- creating a stargazer namespace
- creating stargazer service accounts and roles

Signed-off-by: Sam Hague <shague@gmail.com>