Adds UNP CRD and controller (#13)
* Adds UNP CRD and controller

With this patch the architecture is reworked so that a controller is
generic and one is spawned per specified type in the config. Each
controller has different event "handler" drivers which are specific to
the type of controller being used. The only handler which exists with
this patch is UNP. This patch allows UNP to be automatically registered
as a CRD with k8s, and create/delete actions for instances of UNP will
be read and written into Nimbess etcd.

Other changes include:
 - Consolidating and cleaning up k8s deployment files
 - Modifying etcd port to be 52379 as 12379 is used by skydive

Todo:
 - Add update support in handler for UNP
 - Update UNP CRD with a more realistic version

Signed-off-by: Tim Rozet <trozet@redhat.com>